refactor(backend): use vite_embed crate for development
Allows hot-reloading of Vite files, also uses Just for building instead of build.rs
