+++ frontend/index.html
@@ -6,11 +6,12 @@
   <link rel="icon" href="/favicon.png" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>DietPi Dashboard</title>
+  <!--vite-embed css injection-->
 </head>
 
 <body>
   <div id="app"></div>
-  <script type="module" src="/src/main.ts"></script>
+  <!--vite-embed script injection-->
 </body>
 
 </html>
